refactor: rename service to articles, renamed classes and update pom.xml
