+++ src/main/java/org/acme/Article.java
@@ -1,10 +1,10 @@
-package model;
+package org.acme;
 
 import io.quarkus.hibernate.orm.panache.PanacheEntity;
 import jakarta.persistence.Entity;
 
 @Entity
-public class Product extends PanacheEntity {
+public class Article extends PanacheEntity {
     public String itemId;
     public String name;
     public String description;
