+++ src/main/java/org/acme/DataImport.java
@@ -1,4 +1,4 @@
-package service;
+package org.acme;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -6,7 +6,6 @@ import io.quarkus.runtime.StartupEvent;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.enterprise.event.Observes;
 import jakarta.transaction.Transactional;
-import model.Product;
 
 import java.io.InputStream;
 import java.util.List;
@@ -16,21 +15,21 @@ public class DataImport {
 
     @Transactional
     public void loadInitialData(@Observes StartupEvent ev){
-        if(Product.count() > 0){
+        if(Article.count() > 0){
             return;
         }
 
         try {
             ObjectMapper mapper = new ObjectMapper();
-            TypeReference<List<Product>> typeReference = new TypeReference<List<Product>>(){};
+            TypeReference<List<Article>> typeReference = new TypeReference<List<Article>>(){};
 
             // Datei aus resources laden
             InputStream inputStream = Thread.currentThread().getContextClassLoader().getResourceAsStream("products.json");
-            List<Product> products = mapper.readValue(inputStream, typeReference);
+            List<Article> articles = mapper.readValue(inputStream, typeReference);
 
 
             // In die DB speichern
-            Product.persist(products);
+            Article.persist(articles);
         } catch (Exception e) {
             System.out.println("Fehler beim JSON-Import: " + e.getMessage());
         }
