--- src/main/java/org/acme/MyLivenessCheck.java
@@ -1,15 +1,0 @@
-package org.acme;
-
-import org.eclipse.microprofile.health.HealthCheck;
-import org.eclipse.microprofile.health.HealthCheckResponse;
-import org.eclipse.microprofile.health.Liveness;
-
-@Liveness
-public class MyLivenessCheck implements HealthCheck {
-
-    @Override
-    public HealthCheckResponse call() {
-        return HealthCheckResponse.up("alive");
-    }
-
-}
