+++ src/main/resources/application.properties
@@ -1,13 +1,10 @@
 # Datenbank-Typ und Ort
-quarkus.datasource.db-kind=h2
-quarkus.datasource.jdbc.url=jdbc:h2:mem:default
-quarkus.datasource.username=sa
-quarkus.datasource.password=
+quarkus.datasource.db-kind=postgresql
 
 # Hibernate Einstellungen
 quarkus.hibernate-orm.database.generation=drop-and-create
 quarkus.hibernate-orm.log.sql=true
 
-# OIDC Sicherheit (damit die Endpunkte wissen, dass sie prüfen müssen)
-quarkus.http.auth.permission.public.paths=/products,/products/*
+# OIDC Sicherheit
+quarkus.http.auth.permission.public.paths=/articles,/articles/*
 quarkus.http.auth.permission.public.policy=permit
